Relevant BUGIDs: rhbz#460241
Purpose of commit: bugfix

Commit summary:
---------------
2008-08-29  Tomas Mraz <t8m@centrum.cz>

        * modules/pam_loginuid/pam_loginuid.c(set_loginuid): Uids
        are unsigned.

+++ modules/pam_loginuid/pam_loginuid.c
@@ -53,7 +53,7 @@ static int set_loginuid(pam_handle_t *pamh, uid_t uid)
 	int fd, count, rc = 0;
 	char loginuid[24];
 
-	count = snprintf(loginuid, sizeof(loginuid), "%d", uid);
+	count = snprintf(loginuid, sizeof(loginuid), "%lu", (unsigned long)uid);
 	fd = open("/proc/self/loginuid", O_NOFOLLOW|O_WRONLY|O_TRUNC);
 	if (fd < 0) {
 		if (errno != ENOENT) {
